Add files via upload
This file contains the main program which includes reading with the use of a Terror Tuple class and manipulating character data using Lemma Tuple class and performing one-hot encoding to make it suitable for our model i.e. Logistic Regression. Developing and defining the data pipeline and finally applying the model using SparkML.